+++ play/src/front/WebRtc/MediaManager.ts
@@ -38,14 +38,7 @@ export class MediaManager {
     private canSendNotification = true;
     private canPlayNotificationMessage = true;
 
-    private _faviconsHref: Map<string, string> = new Map();
-
     constructor() {
-        const links = document.querySelectorAll("link[rel~='icon']");
-        for (const link of links) {
-            this._faviconsHref.set(link.getAttribute("sizes") as string, link.href);
-        }
-
         localeDetector()
             .catch(() => {
                 throw new Error("Cannot load locale on media manager");
@@ -217,8 +210,6 @@ export class MediaManager {
     }
 
     public createNotification(userName: string, notificationType: NotificationType, forum: string | null = null) {
-        this.pushNotificationFavicon();
-
         if (document.hasFocus()) {
             return;
         }
@@ -259,70 +250,6 @@ export class MediaManager {
                     return setTimeout(() => (this.canPlayNotificationMessage = true), TIME_NOTIFYING_MILLISECOND);
                 })
                 .catch((err) => console.error("Trying to play notification message error: ", err));
-        }
-    }
-
-    public pushNotificationFavicon() {
-        // Get list of favicon
-        const links = document.querySelectorAll("link[rel~='icon']");
-        for (const link of links) {
-            // Load favicon image
-            const img = document.createElement("img");
-            img.src = link.href;
-            img.onload = () => {
-                const faviconSize = 16;
-                // Cretae image notification for fiveicon
-                const canvas = document.createElement("canvas");
-                canvas.width = faviconSize;
-                canvas.height = faviconSize;
-                const ctx = canvas.getContext("2d") as CanvasRenderingContext2D;
-
-                // Draw Original Favicon as Background
-                ctx.drawImage(img, 0, 0, faviconSize, faviconSize);
-
-                // Draw Notification Circle
-                ctx.beginPath();
-                ctx.arc(canvas.width - faviconSize / 3, faviconSize / 3, faviconSize / 3, 0, 2 * Math.PI);
-                ctx.fillStyle = "#FF0000";
-                ctx.fill();
-
-                // Replace favicon
-                link.href = canvas.toDataURL();
-            };
-        }
-    }
-
-    public pushOriginalFavicon() {
-        // Get list of favicon and remove it
-        const links = document.querySelectorAll("link[rel~='icon']");
-        for (const link of links) {
-            link.remove();
-        }
-
-        const head = document.querySelector("head") as HTMLHeadElement;
-        for (const sizes of this._faviconsHref.keys()) {
-            // Load previous favicon image
-            const img = document.createElement("img");
-            img.src = this._faviconsHref.get(sizes) as string;
-            img.onload = () => {
-                const faviconSize = 16;
-                // Cretae image notification for fiveicon
-                const canvas = document.createElement("canvas");
-                canvas.width = faviconSize;
-                canvas.height = faviconSize;
-                const ctx = canvas.getContext("2d") as CanvasRenderingContext2D;
-
-                // Draw Original Favicon as Background
-                ctx.drawImage(img, 0, 0, faviconSize, faviconSize);
-
-                // Replace favicon
-                const link = document.createElement("link");
-                link.href = canvas.toDataURL();
-                link.setAttribute("sizes", sizes);
-                link.type = "image/png";
-                link.rel = "icon";
-                head.appendChild(link);
-            };
         }
     }
 }
